Renaming the duplicate detection folder
