+++ freediscovery/clustering/__init__.py
@@ -6,4 +6,6 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from .base import (select_top_words, _generate_lsi, ClusterLabels, Clustering)
-from .dendogram import _DendrogramChildren
+
+from .dendrogram import _DendrogramChildren
+
